requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
